+++ pywa/types/__init__.py
@@ -50,6 +50,7 @@ from .others import (
     Pagination,
     StorageConfiguration,
     UserIdentityChangeSettings,
+    Unsupported,
 )
 from .templates import (
     Template,